+++ main/src/me/ricky/kbot/core/util/Collect.kt
@@ -1,3 +1,4 @@
 package me.ricky.kbot.core.util
 
-fun <K, V> MutableMap<K, V>.getOrSet(id: K, value: V) = getOrPut(id, { value })
+fun <K, V> MutableMap<K, V>.getOrSet(key: K, value: V) = getOrPut(key, { value })
+fun <K, V> Map<K, V>.notContainsKey(key: K) = !containsKey(key)